axios call in app.js
