+++ src/components/Book.js
@@ -1,22 +1,7 @@
-import axios from "axios";
-import React, { useState, useEffect } from "react";
-
 function Book(props) {
-  const [books, setBooks] = useState([]);
-
-  useEffect(() => {
-    getData();
-  }, []);
-
-  function getData() {
-    axios.get("http://localhost:8000/books").then((res) => {
-      const bookData = res.data;
-      setBooks(bookData);
-    });
-  }
   return (
     <div>
-      {books.map((book) => {
+      {props.books.map((book) => {
         return (
           <div className="book" key={book._id}>
             <img src={book.image} alt={book.title} className="book-image" />
